src/pegasus/agents/src/python_failed_reorg_20250828_145457/bridge_health_monitor.py
#!/usr/bin/env python3
# Bridge System Monitor
import json
import os
import sys
import time

# Add correct paths
sys.path.append('${CLAUDE_PROJECT_ROOT:-$(dirname "$0")/../../}agents')
sys.path.append('${CLAUDE_AGENTS_ROOT:-$(dirname "$0")}/../03-BRIDGES')

from agent_bridge_main import bridge, task_agent_invoke
from claude_agent_bridge import AgentConfig, BinaryBridgeConnection


class BridgeMonitor:
    def __init__(self):
        self.metrics = {}
        self.start_time = time.time()

    def log_metric(self, metric_name, value):
        timestamp = time.time()
        if metric_name not in self.metrics:
            self.metrics[metric_name] = []
        self.metrics[metric_name].append({"timestamp": timestamp, "value": value})

    def get_summary(self):
        return {
            "uptime": time.time() - self.start_time,
            "total_metrics": len(self.metrics),
            "status": "healthy",
        }


monitor = BridgeMonitor()
