src/pegasus/agents/src/python_failed_reorg_20250828_145457/requirements_production.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asyncio==4.0.0
attrs==25.3.0
bandit==1.8.6
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.5
Cython==3.1.3
flake8==7.3.0
frozenlist==1.7.0
grafana-api==1.0.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.4
mypy==1.17.1
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
opentelemetry-api==1.36.0
opentelemetry-exporter-prometheus==0.57b0
opentelemetry-instrumentation==0.57b0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.22.1
propcache==0.3.2
psutil==7.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
stevedore==5.5.0
structlog==25.4.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvloop==0.21.0
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
